fix: Use FP16 dtype for all_embeddings tensor
Changed torch.zeros() calls to use dtype=torch.float16 to match model
output precision. This fixes the error:
"Index put requires the source and destination dtypes match"

Modified files:
- executor.py: all_embeddings tensor
- pipeline_executor.py: final_tensor in _collect_final_embeddings

+++ executer/executor.py
@@ -170,10 +170,10 @@ class PipelineExecutor:
         print(f"Execution Mode: {'Pipeline Parallel' if use_pipeline_parallelism else 'Sequential'}")
         print(f"{'='*70}\n")
 
-        # 准备输出
+        # 准备输出（使用 FP16 以匹配模型输出精度）
         total_nodes = self.data_loader.full_data.num_nodes
         hidden_dim = 256  # GraphSAGE默认hidden dimension
-        all_embeddings = torch.zeros(total_nodes, hidden_dim)
+        all_embeddings = torch.zeros(total_nodes, hidden_dim, dtype=torch.float16)
 
         per_subgraph_times = []
         per_cluster_times = []
